services/user-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3
psycopg2-binary==2.9.9
redis==5.0.1
pika==1.3.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
prometheus-client==0.19.0
python-json-logger==2.0.7